pyproject.toml
[tool.poetry]
name = "fairy-tales"
version = "0.1.0"
description = ""
authors = ["A.Sidorova"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.0"
numpy = "^1.26.4"
tqdm = "^4.66.2"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
hydra-core = "^1.3.2"
aiogram = "^3.4.1"


[tool.poetry.group.dev.dependencies]
selenium = "^4.17.2"
requests = "^2.31.0"
pre-commit = "^3.6.1"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
dvc = "^3.44.0"
dvc-gdrive = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
